fixing build

+++ belokurskaya.sofia/S2/stack.hpp
@@ -100,8 +100,8 @@ namespace belokurskaya
     private:
       T* data_;
       size_t size_;
-      size_t top_;
-      size_t capacity_;
+      int top_;
+      int capacity_;
       const size_t initial_capacity_ = 3;
       const size_t capacity_change_factor_ = 2;
